+++ modules/admin/src/models/UserOnline.php
@@ -11,6 +11,9 @@ use yii\db\ActiveRecord;
  * @property integer $user_id
  * @property integer $last_timestamp
  * @property string $invoken_route
+ * @property string $lock_pk
+ * @property string $lock_table
+ * @property string $lock_description
  */
 final class UserOnline extends ActiveRecord
 {
@@ -30,7 +33,9 @@ final class UserOnline extends ActiveRecord
         return [
             [['user_id', 'last_timestamp', 'invoken_route'], 'required'],
             [['user_id', 'last_timestamp'], 'integer'],
+            [['lock_table', 'lock_description'], 'string'],
             [['invoken_route'], 'string', 'max' => 120],
+            [['lock_pk'], 'safe'],
         ];
     }
     
@@ -44,31 +49,72 @@ final class UserOnline extends ActiveRecord
             'user_id' => 'User ID',
             'last_timestamp' => 'Last Timestamp',
             'invoken_route' => 'Invoken Route',
+            'lockroute' => 'Lockroute',
         ];
     }
-
-    public static function refreshUser($userId, $route)
+    
+    /**
+     * @return \yii\db\ActiveQuery
+     */
+    public function getUser()
     {
-        $time = time();
-        $model = self::find()->where(['user_id' => $userId])->one();
-        // exists $this->user_id in table?
-        if (!$model) {
-            // insert
-            $model = new self();
-            $model->last_timestamp = $time;
-            $model->user_id = $userId;
-            $model->invoken_route = $route;
-            $model->insert();
-        } else {
-            $model->last_timestamp = $time;
-            $model->invoken_route = $route;
-            $model->update();
+        return $this->hasOne(User::class, ['id' => 'user_id']);
+    }
+    
+    // static methods
+
+    public static function lock($userId, $table, $pk, $description)
+    {
+        $model = self::findOne(['user_id' => $userId]);
+        
+        if ($model) {
+            $model->updateAttributes([
+                'last_timestamp' => time(),
+                'lock_table' => $table,
+                'lock_pk' => $pk,
+                'lock_description' => $description,
+            ]);
         }
     }
+    
+    public static function unlock($userId)
+    {
+        $model = self::findOne(['user_id' => $userId]);
+    
+        if ($model) {
+            $model->updateAttributes([
+                'last_timestamp' => time(),
+                'lock_table' => '',
+                'lock_pk' => '',
+                'lock_description' => '',
+            ]);
+        }
+    }
+    
+    /**
+     * Refresh the state of the current user, or add if not exists.
+     * 
+     * @param integer $userId
+     * @param string $route
+     */
+    public static function refreshUser($userId, $route)
+    {
+        $model = self::findOne(['user_id' => $userId]);
+        if ($model) {
+            return (bool) $model->updateAttributes(['last_timestamp' => time(), 'invoken_route' => $route]);
+        }
+        
+        $model = new self(['last_timestamp' => time(), 'user_id' => $userId, 'invoken_route' => $route]);
+        return $model->save();
+    }
 
+    /**
+     * Remove the given user id if exists.
+     * @param unknown $userId
+     */
     public static function removeUser($userId)
     {
-        $model = self::find()->where(['user_id' => $userId])->one();
+        $model = self::findOne(['user_id' => $userId]);
         if ($model) {
             $model->delete();
         }
@@ -85,33 +131,43 @@ final class UserOnline extends ActiveRecord
             $model->delete();
         }
     }
-
+    
+    /**
+     * Get the user online list.
+     * 
+     * @return array
+     */
     public static function getList()
     {
         $time = time();
         $return = [];
-        foreach (self::find()->asArray()->all() as $item) {
-            $user = User::findOne($item['user_id']);
-            if ($user) {
-                $inactiveSince = $time - $item['last_timestamp'];
-                $return[] = [
-                    'firstname' => $user->firstname,
-                    'lastname' => $user->lastname,
-                    'email' => $user->email,
-                    'last_timestamp' => $item['last_timestamp'],
-                    'is_active' => ($inactiveSince >= 120) ? false : true,
-                    'inactive_since' => round(($inactiveSince / 60)).' min',
-                ];
-            }
+        
+        foreach (self::find()->with('user')->all() as $item) {
+            /* @var $item \luya\admin\models\UserOnline */
+            $inactiveSince = ($time - $item->last_timestamp);
+            $return[] = [
+                'firstname' => $item->user->firstname,
+                'lastname' => $item->user->lastname,
+                'email' => $item->user->email,
+                'last_timestamp' => $item->last_timestamp,
+                'is_active' => ($inactiveSince >= 120) ? false : true,
+                'inactive_since' => round(($inactiveSince / 60)).' min',
+                'lock_description' => $item->lock_description,
+            ];
         }
 
         return $return;
     }
 
+    /**
+     * Get the number uf online users.
+     * 
+     * @return integer
+     */
     public static function getCount()
     {
         static::clearList();
 
-        return count(self::find()->asArray()->all());
+        return self::find()->count();
     }
 }