+++ modules/admin/src/ngrest/render/RenderCrud.php
@@ -8,7 +8,6 @@ use luya\admin\components\Auth;
 use luya\admin\models\Lang;
 use luya\admin\ngrest\NgRest;
 use luya\admin\ngrest\base\Render;
-use yii\helpers\Json;
 use yii\base\InvalidConfigException;
 
 /**